+++ src/components/Header/Header.jsx
@@ -1,12 +1,10 @@
 import * as S from './styles';
 import logo from 'assets/img/logo.png';
-import { Search } from '..';
-import { Header__Sidebar } from '..';
-import { useNavigate } from 'react-router-dom';
+import { Search } from 'components';
+import { UserInfo } from 'components';
 import { Link } from 'react-router-dom';
 
 export const Header = () => {
-  const navigate = useNavigate();
   return (
     <S.Header>
       <S.LogoContainer>
@@ -15,7 +13,7 @@ export const Header = () => {
         </Link>
       </S.LogoContainer>
       <Search />
-      <Header__Sidebar />
+      <UserInfo />
     </S.Header>
   );
 };
