+++ src/components/MusicPlayer/components/ProgressBar/styles.js
