+++ src/components/index.js
@@ -1,8 +1,8 @@
-// const { Header__Sidebar } = require("./Header__Sidebar/Header__Sidebar");
-
-export { Header__Sidebar } from './Header__Sidebar/Header__Sidebar';
+// Header
+export { UserInfo } from './UserInfo/UserInfo';
 export { Search } from './Search/Search';
 export { Header } from './Header/Header';
+// Login
 export { EntryBtn } from './EntryBtn/EntryBtn';
 export { EntryInput } from './EntryInput/EntryInput';
 export { RegBtn } from './RegBtn/RegBtn';
@@ -12,7 +12,9 @@ export { Nav } from './Nav/Nav';
 // Music Player
 export { Controls } from './MusicPlayer/components/Controls/Controls';
 export { LikeDislike } from './MusicPlayer/components/LikeDislike/LikeDislike';
-export { MusicPlayer__ProgressBar } from './MusicPlayer/components/MusicPlayer__ProgressBar/MusicPlayer__ProgressBar';
+export { ProgressBar } from './MusicPlayer/components/ProgressBar/ProgressBar';
 export { TrackInfo } from './MusicPlayer/components/TrackInfo/TrackInfo';
 export { Volume } from './MusicPlayer/components/Volume/Volume';
-export { MusicPlayer__Interface } from './MusicPlayer/components/MusicPlayer__Interface/MusicPlayer__Interface';
+export { Interface } from './MusicPlayer/components/Interface/Interface';
+// Sidebar
+export { Sidebar__Item } from './Sidebar__Item/Sidebar__Item';
