+++ src/pages/Tracks/Tracks.jsx
@@ -1,6 +1,6 @@
 import * as S from './styles';
-import { Tracks__SearchFilter } from '../Tracks__SearchFilter/Tracks__SearchFilter';
-import { Tracks__List } from '../Tracks__List/Tracks__List';
+import { Tracks__SearchFilter } from '../MainPage/components/Tracks__SearchFilter/Tracks__SearchFilter';
+import { Tracks__List } from '../MainPage/components/Tracks__List/Tracks__List';
 import { useEffect, useState } from 'react';
 import { setTracks } from 'store/tracksSlice';
 import { fetchAllTracks } from 'helpers/DAL';
@@ -11,11 +11,15 @@ export const Tracks = ({ heading = 'Треки', playlist }) => {
   const dispatch = useDispatch();
   // Загружаю все треки
   useEffect(() => {
-    fetchAllTracks().then((data) => {
-      dispatch(setTracks(data));
-      setLoadingClass('');
-    }).catch(error => {console.warn(error.message)
-      dispatch(setTracks(error.message))});
+    fetchAllTracks()
+      .then((data) => {
+        dispatch(setTracks(data));
+        setLoadingClass('');
+      })
+      .catch((error) => {
+        console.warn(error.message);
+        dispatch(setTracks(error.message));
+      });
   }, []);
 
   return (
